language_assistant/language_assistant/requirements.txt
openai
streamlit
gtts
bcrypt
gtts
plotly
streamlit-audiorecorder
scikit-learn
